+++ data/01_raw/tangaras.csv
@@ -1,14 +1,14 @@
 ID,GEOHASH,MAC,GEOLOCATION,LATITUDE,LONGITUDE,DATETIME
-TANGARA_2BBA,d29e6b4,D29ESP32DE02BBA,3.3844757080078125 -76.51634216308594,3.3844757080078125,-76.51634216308594,2022-09-14 01:26:23.571277-05:00
-TANGARA_14D6,d29dfx4,D29ESP32DED14D6,3.3350372314453125 -76.52732849121094,3.3350372314453125,-76.52732849121094,2022-09-14 01:26:23.571277-05:00
-TANGARA_1CE2,d29e4cv,D29ESP32DED1CE2,3.3501434326171875 -76.51222229003906,3.3501434326171875,-76.51222229003906,2022-09-14 01:26:23.571277-05:00
-TANGARA_1FCA,d29e48s,D29ESP32DED1FCA,3.3432769775390625 -76.52458190917969,3.3432769775390625,-76.52458190917969,2022-09-14 01:26:23.571277-05:00
-TANGARA_2492,d29e64g,D29ESP32DED2492,3.3995819091796875 -76.54792785644531,3.3995819091796875,-76.54792785644531,2022-09-14 01:26:23.571277-05:00
-TANGARA_2FF6,d29e66v,D29ESP32DED2FF6,3.3995819091796875 -76.53419494628906,3.3995819091796875,-76.53419494628906,2022-09-14 01:26:23.571277-05:00
-TANGARA_48C6,d29ee19,D29TTGOT7D48C6,3.4366607666015625 -76.50672912597656,3.4366607666015625,-76.50672912597656,2022-09-14 01:26:23.571277-05:00
-TANGARA_4D7A,d29ed5r,D29TTGOT7D4D7A,3.4462738037109375 -76.54243469238281,3.4462738037109375,-76.54243469238281,2022-09-14 01:26:23.571277-05:00
-TANGARA_532E,d29ed74,D29TTGOT7D532E,3.4449005126953125 -76.53831481933594,3.4449005126953125,-76.53831481933594,2022-09-14 01:26:23.571277-05:00
-TANGARA_EA06,d29ed5r,D29TTGOTD8EA06,3.4462738037109375 -76.54243469238281,3.4462738037109375,-76.54243469238281,2022-09-14 01:26:23.571277-05:00
-TANGARA_F1AE,d29eg66,D29TTGOTD8F1AE,3.4847259521484375 -76.49436950683594,3.4847259521484375,-76.49436950683594,2022-09-14 01:26:23.571277-05:00
-TANGARA_FAC6,d29ed5r,D29TTGOTD8FAC6,3.4462738037109375 -76.54243469238281,3.4462738037109375,-76.54243469238281,2022-09-14 01:26:23.571277-05:00
-TANGARA_06BE,d29e6de,D29TTGOTD906BE,3.3982086181640625 -76.52595520019531,3.3982086181640625,-76.52595520019531,2022-09-14 01:26:23.571277-05:00
+TANGARA_2BBA,d29e6b4,D29ESP32DE02BBA,3.3844757080078125 -76.51634216308594,3.3844757080078125,-76.51634216308594,2022-09-14 14:14:21.658719-05:00
+TANGARA_14D6,d29dfx4,D29ESP32DED14D6,3.3350372314453125 -76.52732849121094,3.3350372314453125,-76.52732849121094,2022-09-14 14:14:21.658719-05:00
+TANGARA_1CE2,d29e4cv,D29ESP32DED1CE2,3.3501434326171875 -76.51222229003906,3.3501434326171875,-76.51222229003906,2022-09-14 14:14:21.658719-05:00
+TANGARA_1FCA,d29e48s,D29ESP32DED1FCA,3.3432769775390625 -76.52458190917969,3.3432769775390625,-76.52458190917969,2022-09-14 14:14:21.658719-05:00
+TANGARA_2492,d29e64g,D29ESP32DED2492,3.3995819091796875 -76.54792785644531,3.3995819091796875,-76.54792785644531,2022-09-14 14:14:21.658719-05:00
+TANGARA_2FF6,d29e66v,D29ESP32DED2FF6,3.3995819091796875 -76.53419494628906,3.3995819091796875,-76.53419494628906,2022-09-14 14:14:21.658719-05:00
+TANGARA_48C6,d29ee19,D29TTGOT7D48C6,3.4366607666015625 -76.50672912597656,3.4366607666015625,-76.50672912597656,2022-09-14 14:14:21.658719-05:00
+TANGARA_4D7A,d29ed5r,D29TTGOT7D4D7A,3.4462738037109375 -76.54243469238281,3.4462738037109375,-76.54243469238281,2022-09-14 14:14:21.658719-05:00
+TANGARA_532E,d29ed74,D29TTGOT7D532E,3.4449005126953125 -76.53831481933594,3.4449005126953125,-76.53831481933594,2022-09-14 14:14:21.658719-05:00
+TANGARA_EA06,d29ed5r,D29TTGOTD8EA06,3.4462738037109375 -76.54243469238281,3.4462738037109375,-76.54243469238281,2022-09-14 14:14:21.658719-05:00
+TANGARA_F1AE,d29eg66,D29TTGOTD8F1AE,3.4847259521484375 -76.49436950683594,3.4847259521484375,-76.49436950683594,2022-09-14 14:14:21.658719-05:00
+TANGARA_FAC6,d29ed5r,D29TTGOTD8FAC6,3.4462738037109375 -76.54243469238281,3.4462738037109375,-76.54243469238281,2022-09-14 14:14:21.658719-05:00
+TANGARA_06BE,d29e6de,D29TTGOTD906BE,3.3982086181640625 -76.52595520019531,3.3982086181640625,-76.52595520019531,2022-09-14 14:14:21.658719-05:00
