+++ src/tangara_pipeline/pipeline_registry.py
@@ -3,6 +3,7 @@ from typing import Dict
 
 from kedro.pipeline import Pipeline, pipeline
 
+from tangara_pipeline.pipelines import pm25
 from tangara_pipeline.pipelines import pm25_nowcast_aqi
 
 
@@ -12,9 +13,11 @@ def register_pipelines() -> Dict[str, Pipeline]:
     Returns:
         A mapping from a pipeline name to a ``Pipeline`` object.
     """
+    pm25_pipeline = pm25.create_pipeline()
     pm25_nowcast_aqi_pipeline = pm25_nowcast_aqi.create_pipeline()
 
     return {
-        "__default__": pm25_nowcast_aqi_pipeline,
+        "__default__": pm25_pipeline,
+        "pm25": pm25_pipeline,
         "pm25_nowcast_aqi": pm25_nowcast_aqi_pipeline
     }
